+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fixed automatic driver updates not working when the leader instance is removed
+  from the project
+
 ### Added
 
 - Initial release of Device Programmer